Rename imported files
