+++ __tests__/unit/repository/medicalequipment/add-medicalequipment.repository.unit.test.tsx
@@ -1,4 +1,4 @@
-import MedicalEquipmentRepository from "../../../../src/repository/add-medicalequipment.repository";
+import MedicalEquipmentRepository from "../../../../src/repository/medicalequipment.repository";
 import {
   AddMedicalEquipmentDTO,
   AddMedicalEquipmentResponseDTO,
